+++ packages/kernel/CHANGELOG.md
@@ -9,11 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Actor.snap() sets properties without smoothing
 - Actor.set() automatically calls this.propertySet() if the actor has that method.
 - ScaleTo, RotateTo & TranslateTo deprecated
+- AM_Smoothed deprecated (Use AM_Spatial)
 - AM_Spatial.local setter added
 - m4_getScale(), m4_getScaleRotationTranslation() added
 
 ### Changed
 - Pawn destruction handled by PawnManager, not Pawn itself
+- Services no longer have star-up options
 
 ### Fixed
 - m4_getRotation() correctly ignores scale
@@ -25,7 +27,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ObjectCache
 - View Asset Cache
 - Generic RenderManager
-
 
 ## [1.4.0] - 2022-10-3
 ### Added